Add an argument to debmarshal.distros.base.captureCall for passing a string on stdin.
Because subprocess really just wraps os.fork, os.dup2, etc., the
stdin, stdout, and stderr arguments must be bona fide fds, not just
file-like objects, so you can't use things like StringIOs.

If you want to be able to pass in a string instead of an fd, you need
to specify the argument to Popen.communicate.

Signed-off-by: Evan Broder <ebroder@google.com>
[dld: Python has no IO wrapper for stings like Perl's IO::String?]
[ebroder: There's StringIO, but the stdin argument has to be an actual
process-level fd, not an fd-like-Python-object]