Add a callWait method to privops for "asynchronous" method calls.
This is for privileged operations that take too long to run in a
method call without timing out. For instance, creating a custom disk
image.

callWait will make the call normally (which should return quickly),
then wait for the privileged daemon to call a method provided by the
unprivileged script.

Signed-off-by: Evan Broder <ebroder@google.com>