Filter out tests known to be slow.
The test suite is beginning to get a bit unwieldy, so any test that's
"tagged" with the slow attribute won't be run by default.

All tests can be run by running nosetests with -a slow (which enables
slow tests in addition to !slow tests).

Unfortunately, this doesn't work with test classes inheriting from
MoxTestBase because the metaclass won't preserve arbitrary attributes
on methods, but tests that are using mox are almost certainly snappy
to begin with.

Signed-off-by: Evan Broder <ebroder@google.com>